refactor: separate Role Filter into its own slice and directory
- Created roleFilterSlice.ts to isolate Role Filter state and actions.
- Moved and refactored Role Filter components into src/feature/rolefilter/.
- Decomposed large components into focused smaller ones (RoleSearchInput, RoleGrid, RoleFilterCardLayout, RoleFilterAddButton).
- Improved dialog system to use union types and encapsulated state.
- Updated all imports and references across the project.
- Verified with unit tests, E2E tests, and static checks.
