+++ src/slices/exrOptionViewerSlice.ts
@@ -8,7 +8,6 @@ import {
 import type {
 	ExROptionValueData,
 	ExRTabId,
-	RoleAssignFilterSetUI,
 	UniqueOptionId,
 	UpdatedOptions,
 } from "../type";
@@ -26,7 +25,6 @@ export interface ExROptionViewerSlice {
 	exrValue: Record<UniqueOptionId, ExROptionValueData>;
 	isExROptionActive: Record<UniqueOptionId, boolean>;
 	highlightedExROptionId: UniqueOptionId | null;
-	roleFilterSet: Record<string, RoleAssignFilterSetUI>;
 	setSelectedExRTabId: (id: ExRTabId) => void;
 	setIsExRTabPending: (isPending: boolean) => void;
 	toggleExRCategory: (categoryId: number) => void;
@@ -40,11 +38,6 @@ export interface ExROptionViewerSlice {
 		valueData: Record<UniqueOptionId, ExROptionValueData>,
 		optionActiveData: Record<UniqueOptionId, boolean>,
 	) => void;
-	setRoleFilterSet: (data: Record<string, RoleAssignFilterSetUI>) => void;
-	addRoleFilter: (guid: string) => void;
-	deleteRoleFilter: (guid: string) => void;
-	addRoleToFilter: (guid: string, roleId: number, roleName: string) => void;
-	removeRoleFromFilter: (guid: string, roleId: number) => void;
 	validateOpenedIds: () => void;
 }
 
@@ -62,7 +55,6 @@ export const createExROptionViewerSlice: StateCreator<ExROptionViewerSlice> = (
 		exrValue: {},
 		isExROptionActive: {},
 		highlightedExROptionId: null,
-		roleFilterSet: {},
 		presetNames: loadPresetNamesFromLocalStorage(),
 		isPresetDropdownOpen: false,
 		setSelectedExRTabId: (id: ExRTabId) => {
@@ -158,68 +150,6 @@ export const createExROptionViewerSlice: StateCreator<ExROptionViewerSlice> = (
 			set({
 				exrValue: valueData,
 				isExROptionActive: optionActiveData,
-			});
-		},
-		setRoleFilterSet: (data: Record<string, RoleAssignFilterSetUI>) => {
-			set({ roleFilterSet: data });
-		},
-		addRoleFilter: (guid: string) => {
-			set((state) => {
-				return {
-					roleFilterSet: {
-						...state.roleFilterSet,
-						[guid]: {
-							AssignNum: 0,
-							Roles: [],
-						},
-					},
-				};
-			});
-		},
-		deleteRoleFilter: (guid: string) => {
-			set((state) => {
-				const nextRoleFilterSet = { ...state.roleFilterSet };
-				delete nextRoleFilterSet[guid];
-				return {
-					roleFilterSet: nextRoleFilterSet,
-				};
-			});
-		},
-		addRoleToFilter: (guid: string, roleId: number, roleName: string) => {
-			set((state) => {
-				const filter = state.roleFilterSet[guid];
-				if (!filter) {
-					return state;
-				}
-				if (filter.Roles.some((r) => r.id === roleId)) {
-					return state;
-				}
-				return {
-					roleFilterSet: {
-						...state.roleFilterSet,
-						[guid]: {
-							...filter,
-							Roles: [...filter.Roles, { id: roleId, name: roleName }],
-						},
-					},
-				};
-			});
-		},
-		removeRoleFromFilter: (guid: string, roleId: number) => {
-			set((state) => {
-				const filter = state.roleFilterSet[guid];
-				if (!filter) {
-					return state;
-				}
-				return {
-					roleFilterSet: {
-						...state.roleFilterSet,
-						[guid]: {
-							...filter,
-							Roles: filter.Roles.filter((r) => r.id !== roleId),
-						},
-					},
-				};
 			});
 		},
 		validateOpenedIds: () => {
